centralize Color type
