+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/checkbox-list/types.ts
@@ -6,14 +6,7 @@ import { BlockEditProps } from '@wordpress/blocks';
 /**
  * Internal dependencies
  */
-import { FilterBlockContext } from '../../types';
-
-export type Color = {
-	slug: string;
-	class: string;
-	name: string;
-	color: string;
-};
+import type { Color, FilterBlockContext } from '../../types';
 
 export type BlockAttributes = {
 	className: string;
