+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/price-slider/constants.ts
@@ -1,5 +1,1 @@
-export const colorAttributes = [
-	'sliderHandle',
-	'sliderHandleBorder',
-	'slider',
-];
+export const colorNames = [ 'sliderHandle', 'sliderHandleBorder', 'slider' ];
