+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/types.ts
@@ -45,3 +45,10 @@ export type FilterBlockContext = {
 		} >;
 	};
 };
+
+export type Color = {
+	slug: string;
+	class: string;
+	name: string;
+	color: string;
+};
